+++ src/assets/index.go
@@ -3,7 +3,8 @@ package assets
 import (
 	"bytes"
 	"fmt"
-	"math"
+	"strconv"
+	"strings"
 
 	"github.com/mariomac/goblog/src/blog"
 	"github.com/mariomac/goblog/src/visual"
@@ -15,14 +16,34 @@ const (
 
 type IndexGenerator struct {
 	entries   *blog.Entries
+	entriesPerPage int
 	templates *visual.Templater
 }
 
-func (i *IndexGenerator) Get(_ string) (*WebAsset, error) {
-	// TODO: properly paginate entries
+type IndexRender struct {
+	PreviousPage int
+	CurrentPage int
+	NextPage int
+	TotalPages int
+	Entries []*blog.Entry
+}
 
+func (i *IndexGenerator) Get(page string) (*WebAsset, error) {
+	ir := IndexRender{
+		TotalPages: i.entries.Len() / i.entriesPerPage + 1,
+		CurrentPage: 1,
+	}
+	if len(page) != 0 {
+		var err error
+		ir.CurrentPage, err = strconv.Atoi(strings.SplitN(page, "/", 2)[0])
+		if err != nil || ir.CurrentPage <= 0 {
+			return nil, errNotFound{}
+		}
+	}
+	ir.PreviousPage, ir.NextPage = ir.CurrentPage - 1, ir.CurrentPage + 1
+	ir.Entries = i.entries.Sorted(ir.CurrentPage - 1, i.entriesPerPage)
 	body := bytes.Buffer{}
-	if err := i.templates.Render(visual.IndexTemplate, i.entries.Sorted(0, math.MaxInt), &body); err != nil {
+	if err := i.templates.Render(visual.IndexTemplate, ir, &body); err != nil {
 		return nil, internalError{cause: fmt.Errorf("rendering index template: %w", err)}
 	}
 	return &WebAsset{